requirements.txt
# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
mysqlclient==2.2.0
python-decouple==3.8
Pillow==10.1.0
openai==1.3.7
google-cloud-translate==3.12.1
google-cloud-speech==2.21.0
gtts==2.4.0
SpeechRecognition==3.10.0
pydub==0.25.1
requests==2.31.0